add general-dialog-layout for applying workflow styles, refactor image upload component
