+++ projects/styles/lf-laserfiche-lite/sass/_button-lf.scss
@@ -75,10 +75,6 @@ button {
   padding-top: 5px;
 }
 
-.lf-dialog-title {
-  font: 500 20px/32px Roboto,Helvetica Neue,sans-serif;
-}
-
 .lf-dialog-message {
   margin: 10px 0px;
 }
