+++ projects/ui-components/lf-user-feedback/feedback-image-upload/feedback-image-upload.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { FeedbackImageUploadComponent } from './feedback-image-upload.component';
+
+describe('FeedbackImageUploadComponent', () => {
+  let component: FeedbackImageUploadComponent;
+  let fixture: ComponentFixture<FeedbackImageUploadComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ FeedbackImageUploadComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(FeedbackImageUploadComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
