+++ projects/ui-components/lf-user-feedback/feedback-submission/feedback-submission.component.html
@@ -1,9 +1,5 @@
 
-  <div
-  class="user-feedback-content-container lf-dialog-content"
-  (drop)="dropHandler($event)"
-  (dragover)="dragOverHandler($event)"
->
+  <div   class="user-feedback-content-container lf-dialog-content">
   <div class="lf-text-body lf-user-feedback-description">
     {{
       isFeedback
@@ -19,7 +15,7 @@
     rows="12"
     [maxlength]="maxFeedbackLength"
     [(ngModel)]="feedbackTextBox"
-    class="lf-textarea lf-text-body feedback-suggestion-textbox"
+    class="lf-textarea lf-text-placeholder lf-text-body feedback-suggestion-textbox"
     [placeholder]="localizedStrings.PLEASE_DO_NOT_INCLUDE_CONFIDENTIAL_OR_PERSONAL_INFO_IN_FEEDBACK | async"
   ></textarea>
   <span class="lf-text-body character-count">
@@ -27,37 +23,7 @@
     <span> / </span>
     <span>{{ maxFeedbackLength }}</span>
   </span>
-
-  <label for="image-upload" class="lf-text-label">{{ localizedStrings.UPLOAD_IMAGE_OPTIONAL | async }}</label>
-  <div id="image-upload">
-    <div *ngIf="!imageUploaded || !isImageValid" id="drop_zone" class="drop-zone">
-      <label for="upload-button" class="lf-text-body">{{ localizedStrings.DRAG_SELECT_IMAGE | async }}</label>
-      <button
-        id="upload-button"
-        [disabled]="isImageValid"
-        class="lf-button sec-button feedback-file-upload-button"
-        (click)="onInputClickArea()"
-      >
-        {{ localizedStrings.BROWSE | async }}
-      </button>
-    </div>
-
-    <tr *ngIf="isImageValid" class="upload-file-info">
-      <td>{{ imageUploaded?.name }}</td>
-      <td>{{ uploadedImageSize }}</td>
-      <td class="upload-file-remove" (click)="removeImage()">{{ localizedStrings.REMOVE | async }}</td>
-    </tr>
-
-    <input
-      type="file"
-      accept=".jpg,.png,.gif,.webp"
-      class="file-input feedback-file-input"
-      (input)="onFileSelectedAsync($event)"
-      #uploadFile
-    />
-    <p *ngIf="!isImageValid" class="error-message lf-text-body">{{ imageUploadErrorMessage | async }}</p>
-  </div>
-
+  <lib-feedback-image-upload [(feedbackImageBase64)]="feedbackImageBase64"></lib-feedback-image-upload>
   <div class="email-checkbox-container">
     <mat-checkbox
       color="primary"
