+++ projects/ui-components/lf-user-feedback/general-dialog-layout/general-dialog-layout.component.html
@@ -1,0 +1,14 @@
+<div class="wrapper">
+    <div class="header">
+      <ng-content select=".dialog-header"></ng-content>
+    </div>
+  
+    <div class="content-box">
+      <ng-content select=".dialog-content"></ng-content>
+    </div>
+  
+    <div class="footer">
+      <ng-content select=".dialog-footer"></ng-content>
+    </div>
+  </div>
+
