+++ projects/ui-components/lf-user-feedback/general-dialog-layout/general-dialog-layout.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'lib-general-dialog-layout',
+  templateUrl: './general-dialog-layout.component.html',
+  styleUrls: ['./general-dialog-layout.component.css']
+})
+export class GeneralDialogLayoutComponent {}
