+++ projects/ui-components/lf-user-feedback/lf-user-feedback.module.ts
@@ -8,9 +8,11 @@ import { CommonModule } from '@angular/common';
 import { createCustomElement } from '@angular/elements';
 import { FeedbackSuggestionSelectionComponent } from './feedback-suggestion-selection/feedback-suggestion-selection.component';
 import { FeedbackSubmissionComponent } from './feedback-submission/feedback-submission.component';
+import { GeneralDialogLayoutComponent } from './general-dialog-layout/general-dialog-layout.component';
+import { FeedbackImageUploadComponent } from './feedback-image-upload/feedback-image-upload.component';
 
 @NgModule({
-  declarations: [LfUserFeedbackComponent, UserFeedbackDialogComponent, FeedbackSuggestionSelectionComponent, FeedbackSubmissionComponent],
+  declarations: [LfUserFeedbackComponent, UserFeedbackDialogComponent, FeedbackSuggestionSelectionComponent, FeedbackSubmissionComponent, GeneralDialogLayoutComponent, FeedbackImageUploadComponent],
   imports: [
     CommonModule,
     FormsModule,
