+++ projects/ui-components/lf-user-feedback/user-feedback-dialog/user-feedback-dialog.component.css
@@ -1,14 +1,6 @@
-.lf-popup-dialog-title {
-  font-family: 'Open Sans';
-  display: inline;
-  color: #606060;
-  font-weight: 600;
-  margin: 0;
-}
-
 span#pop-up-dialog-close {
-  color: rgb(170, 170, 170);
-  opacity: 0.6;
+  color: #646464;
+  opacity: 1;
 }
 
 span#pop-up-dialog-close:hover {
@@ -17,16 +9,14 @@ span#pop-up-dialog-close:hover {
 
 .lf-close-button {
   cursor: pointer;
-  float: right;
-  display: inline;
-  background-color: inherit;
+  display: flex;
+  height: 24px;
+  width: 24px;
+  align-items: center;
   border: none;
-  padding: 0;
-  margin-right: 1px;
-  margin-left: 2px;
-  margin-top: 4px;
-  line-height: 16px;
-  outline-color: #6D6D6D;
+  background: inherit;
+  justify-content: center;
+  color: #646464;
 }
 
 .lf-textarea {
@@ -64,15 +54,11 @@ span#pop-up-dialog-close:hover {
   margin-top: 2px;
 }
 
-.user-feedback-content-container {
-  padding-right: 30px;
-}
-
 .feedback-suggestion-textbox {
   resize: none;
 }
 
-.lf-dialog-content {
+/* .lf-dialog-content {
   border-bottom-color:rgb(224, 224, 224);
   border-bottom-width: 1px;
   border-bottom-style: solid;
@@ -80,14 +66,10 @@ span#pop-up-dialog-close:hover {
   border-top-width: 1px;
   border-top-style: solid;
   padding-bottom: 20px;
-}
-
-.lf-dialog-footer {
-  padding: 7px 0px;
-}
+} */
 
 .lf-user-feedback-description {
-  margin: 10px 0px;
+  margin-bottom: 10px;
 }
 
 .lf-user-feedback-thank-you {
@@ -181,3 +163,4 @@ span#pop-up-dialog-close:hover {
 button#upload-button {
   margin-left: 10px;
 }
+
